nlp_family_extractor/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
SQLAlchemy>=2.0.35
PyMySQL>=1.1.1
google-genai>=1.0.0
python-dotenv>=1.0.0
